Move mb encoding specific functions to mbyte.
mbyte.c should only contain mb_* functions.

I had to add the include in main.c because of the global function pointers.
Maybe I should put them in mbyte.h.

There is a lot of work to do, I see a lot of `if(enc_utf8)` and co in the
code.
