+++ src/nvim/CMakeLists.txt
@@ -38,6 +38,7 @@ file(MAKE_DIRECTORY ${GENERATED_DIR}/api/private)
 file(MAKE_DIRECTORY ${GENERATED_DIR}/msgpack_rpc)
 file(MAKE_DIRECTORY ${GENERATED_DIR}/tui)
 file(MAKE_DIRECTORY ${GENERATED_DIR}/event)
+file(MAKE_DIRECTORY ${GENERATED_DIR}/mbyte)
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR})
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/os)
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/api)
@@ -45,9 +46,10 @@ file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/api/private)
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/msgpack_rpc)
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/tui)
 file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/event)
+file(MAKE_DIRECTORY ${GENERATED_INCLUDES_DIR}/mbyte)
 
 file(GLOB NEOVIM_SOURCES *.c os/*.c api/*.c api/private/*.c msgpack_rpc/*.c
-  tui/*.c event/*.c)
+  tui/*.c event/*.c mbyte/*.c)
 file(GLOB_RECURSE NEOVIM_HEADERS *.h)
 file(GLOB UNIT_TEST_FIXTURES ${PROJECT_SOURCE_DIR}/test/unit/fixtures/*.c)
